+++ beacon-chain/db/kv/attestations.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/boltdb/bolt"
 	"github.com/gogo/protobuf/proto"
+	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/go-ssz"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db/filters"
 	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
@@ -34,12 +35,11 @@ func (k *Store) Attestations(ctx context.Context, f *filters.QueryFilter) ([]*et
 		// []byte("shard-5"), []byte("parent-root-0x2093923"), etc. to be used for looking up
 		// attestation roots that were stored under each of those indices for O(1) lookup.
 		indices := createIndicesFromFilters(f)
-
 		// Once we have a list of attestation roots that correspond to each
 		// lookup index, we find the intersection across all of them and use
 		// that list of roots to lookup the attestations. These attestations will
 		// meet the filter criteria.
-		keys := sliceutil.TotalIntersectionByteSlices(valuesForIndices(indices, bkt))
+		keys := sliceutil.IntersectionByteSlices(lookupValuesForIndices(indices, bkt)...)
 		for i := 0; i < len(keys); i++ {
 			encoded := bkt.Get(keys[i])
 			att := &ethpb.Attestation{}
@@ -66,6 +66,7 @@ func (k *Store) HasAttestation(ctx context.Context, attRoot [32]byte) bool {
 }
 
 // DeleteAttestation by root.
+// TODO(#3018): Add the ability for batch deletions.
 func (k *Store) DeleteAttestation(ctx context.Context, attRoot [32]byte) error {
 	return k.db.Update(func(tx *bolt.Tx) error {
 		bkt := tx.Bucket(attestationsBucket)
@@ -90,29 +91,12 @@ func (k *Store) SaveAttestation(ctx context.Context, att *ethpb.Attestation) err
 		if bkt.Get(root[:]) != nil {
 			return nil
 		}
-
-		shardKey := append(shardIdx, uint64ToBytes(att.Data.Crosslink.Shard)...)
-		shardRoots := bkt.Get(shardKey)
-		if shardRoots == nil {
-			if err := bkt.Put(shardKey, root[:]); err != nil {
-				return err
-			}
-		} else {
-			if err := bkt.Put(shardKey, append(shardRoots, root[:]...)); err != nil {
-				return err
-			}
+		indices := [][]byte{
+			append(shardIdx, uint64ToBytes(att.Data.Crosslink.Shard)...),
+			append(parentRootIdx, att.Data.Crosslink.ParentRoot...),
 		}
-
-		parentRootKey := append(parentRootIdx, att.Data.Crosslink.ParentRoot...)
-		parentRoots := bkt.Get(parentRootKey)
-		if parentRoots == nil {
-			if err := bkt.Put(parentRootKey, root[:]); err != nil {
-				return err
-			}
-		} else {
-			if err := bkt.Put(parentRootKey, append(parentRoots, root[:]...)); err != nil {
-				return err
-			}
+		if err := updateIndices(indices, root[:], bkt); err != nil {
+			return errors.Wrap(err, "could not update DB indices")
 		}
 		return bkt.Put(root[:], enc)
 	})
@@ -141,29 +125,12 @@ func (k *Store) SaveAttestations(ctx context.Context, atts []*ethpb.Attestation)
 			if bkt.Get(keys[i]) != nil {
 				return nil
 			}
-
-			shardKey := append(shardIdx, uint64ToBytes(atts[i].Data.Crosslink.Shard)...)
-			shardRoots := bkt.Get(shardKey)
-			if shardRoots == nil {
-				if err := bkt.Put(shardKey, keys[i]); err != nil {
-					return err
-				}
-			} else {
-				if err := bkt.Put(shardKey, append(shardRoots, keys[i]...)); err != nil {
-					return err
-				}
+			indices := [][]byte{
+				append(shardIdx, uint64ToBytes(atts[i].Data.Crosslink.Shard)...),
+				append(parentRootIdx, atts[i].Data.Crosslink.ParentRoot...),
 			}
-
-			parentRootKey := append(parentRootIdx, atts[i].Data.Crosslink.ParentRoot...)
-			parentRoots := bkt.Get(parentRootKey)
-			if parentRoots == nil {
-				if err := bkt.Put(parentRootKey, keys[i]); err != nil {
-					return err
-				}
-			} else {
-				if err := bkt.Put(parentRootKey, append(parentRoots, keys[i]...)); err != nil {
-					return err
-				}
+			if err := updateIndices(indices, keys[i], bkt); err != nil {
+				return errors.Wrap(err, "could not update DB indices")
 			}
 			if err := bkt.Put(keys[i], encodedValues[i]); err != nil {
 				return err
@@ -171,33 +138,4 @@ func (k *Store) SaveAttestations(ctx context.Context, atts []*ethpb.Attestation)
 		}
 		return nil
 	})
-}
-
-func createIndicesFromFilters(f *filters.QueryFilter) [][]byte {
-	keys := make([][]byte, 0)
-	for k, v := range f.Filters() {
-		switch k {
-		case filters.Shard:
-			idx := append(shardIdx, uint64ToBytes(v.(uint64))...)
-			keys = append(keys, idx)
-		case filters.ParentRoot:
-			parentRoot := v.([]byte)
-			idx := append(parentRootIdx, parentRoot...)
-			keys = append(keys, idx)
-		}
-	}
-	return keys
-}
-
-func valuesForIndices(indices [][]byte, bkt *bolt.Bucket) [][][]byte {
-	values := make([][][]byte, 0)
-	for _, k := range indices {
-		roots := bkt.Get(k)
-		splitRoots := make([][]byte, 0)
-		for i := 0; i < len(roots); i += 32 {
-			splitRoots = append(splitRoots, roots[i:i+32])
-		}
-		values = append(values, splitRoots)
-	}
-	return values
 }
